Re-enable plugin backwards compatibility tests
Previously, plugin backwards compatibilty was tested against the version
of gpbackup the plugins were released with. However due a change to
example plugin's backup directory structure and the fact that
gpbackup1.7.0 is only compatible with pluginAPI version 0.3.0, the only
way these tests could be re-enabled was if the oldest versions these
tested against was bumped up to when both of these issues were resolved.

Authored-by: Kevin Yeap <kyeap@pivotal.io>